# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User Management System Sample App.",
        "title": "GO \u0026 Gin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/assign-roles": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign multiple roles to a specific user by body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assign role"
                ],
                "summary": "Assign Roles to User",
                "parameters": [
                    {
                        "description": "User ID and Role IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.AssignRolesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/deletebyid_role/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a single role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Delete role by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/get_roles": {
            "get": {
                "description": "Retrieve all registered roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get all roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.GetAllRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/get_users": {
            "get": {
                "description": "Retrieve all registered users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.GetAllUserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/getbyid_role/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get role by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.GetSingleRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/getbyid_user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.GetSingleUserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticate user and return JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login Request",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/register_role": {
            "post": {
                "description": "Registers a role with name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Register a new role",
                "parameters": [
                    {
                        "description": "Role registration data",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.RegisterRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.GetSingleRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Could not register",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/register_user": {
            "post": {
                "description": "Registers a user with email, password, and name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.GetSingleUserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Could not register",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/update_role": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update role body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Update Role",
                "parameters": [
                    {
                        "description": "Role Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/update_user": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update user body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "description": "User Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_api_gin_internal_dtos.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "rest_api_gin_internal_dtos.AssignRolesRequest": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "rest_api_gin_internal_dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.GetAllRoleResponse": {
            "type": "object",
            "properties": {
                "Rolelist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest_api_gin_internal_dtos.GetSingleRoleResponse"
                    }
                }
            }
        },
        "rest_api_gin_internal_dtos.GetAllUserResponse": {
            "type": "object",
            "properties": {
                "Userlist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest_api_gin_internal_dtos.GetSingleUserResponse"
                    }
                }
            }
        },
        "rest_api_gin_internal_dtos.GetSingleRoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.GetSingleUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest_api_gin_internal_dtos.RoleList"
                    }
                }
            }
        },
        "rest_api_gin_internal_dtos.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.RegisterRoleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.RegisterUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "rest_api_gin_internal_dtos.RoleList": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.TokenResponse": {
            "type": "object",
            "properties": {
                "JWTtoken": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.UpdateRoleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "rest_api_gin_internal_dtos.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Provide your access token in the Authorization header using the Bearer scheme.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}